chore: restructure project
1. Extract exchange module to a separated crate
2. Extract engine and utils modules to a separated crate
