+++ app/Cargo.toml
@@ -3,14 +3,11 @@ name = "app"
 version = "0.1.0"
 edition = "2024"
 
+[workspace]
+members = [".", "crates/exchange", "crates/scrapper_engine"]
+
 [dependencies]
-dotenv = "0.15.0"
-tokio-tungstenite = { version = "0.28.0", features = ["rustls-tls-webpki-roots"] }
+exchange = { path = "crates/exchange"}
+scrapper_engine = { path = "crates/scrapper_engine"}
+
 tokio = { version = "1.47.1", features = ["rt", "rt-multi-thread", "macros"] }
-futures-util = "0.3.31"
-rustls = { version = "0.23.32", features = ["ring"] }
-anyhow = "1.0.100"
-serde_json = "1.0.145"
-reqwest = "0.12.23"
-pprof = { version = "0.15.0", features = ["flamegraph"] }
-url = "2.5.7"
