+++ app/crates/exchange/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "exchange"
+version = "0.1.0"
+edition = "2024"
+
+[lib]
+name = "exchange"
+path = "src/lib.rs"
+
+[dependencies]
+tokio-tungstenite = { version = "0.28.0", features = ["rustls-tls-webpki-roots"] }
+tokio = { version = "1.47.1", features = ["rt", "rt-multi-thread", "macros"] }
+futures-util = "0.3.31"
+anyhow = "1.0.100"
+serde_json = "1.0.145"
+reqwest = "0.12.23"
+url = "2.5.7"
