+++ app/crates/exchange/src/structs.rs
