+++ app/crates/scrapper_engine/src/lib.rs
@@ -1,0 +1,3 @@
+mod utils;
+pub mod engine;
+
